Andrey Rohau / Changed AssignTest and StartTest logics Added XML comments for endpoints
